Refactor Scanner related code (#2592)
