+++ core/src/main/java/com/scalar/db/common/FilterableScanner.java
@@ -1,6 +1,5 @@
 package com.scalar.db.common;
 
-import com.google.errorprone.annotations.concurrent.LazyInit;
 import com.scalar.db.api.Result;
 import com.scalar.db.api.Scan;
 import com.scalar.db.api.Scanner;
@@ -11,22 +10,19 @@ import com.scalar.db.util.ScalarDbUtils;
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import java.io.IOException;
 import java.util.ArrayList;
-import java.util.Iterator;
 import java.util.List;
 import java.util.Optional;
 import java.util.Set;
-import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import javax.annotation.concurrent.NotThreadSafe;
 
 @NotThreadSafe
-public class FilterableScanner implements Scanner {
+public class FilterableScanner extends AbstractScanner {
 
   private final Scanner scanner;
   private final List<String> projections;
   private final Set<Conjunction> conjunctions;
   @Nullable private Integer left = null;
-  @LazyInit private ScannerIterator scannerIterator;
 
   public FilterableScanner(Selection selection, Scanner scanner) {
     this.scanner = scanner;
@@ -70,15 +66,6 @@ public class FilterableScanner implements Scanner {
       ret.add(one.get());
     }
     return ret;
-  }
-
-  @Override
-  @Nonnull
-  public Iterator<Result> iterator() {
-    if (scannerIterator == null) {
-      scannerIterator = new ScannerIterator(this);
-    }
-    return scannerIterator;
   }
 
   @Override
