+++ core/src/main/java/com/scalar/db/storage/dynamo/GetItemScanner.java
@@ -1,11 +1,9 @@
 package com.scalar.db.storage.dynamo;
 
 import com.scalar.db.api.Result;
-import com.scalar.db.api.Scanner;
-import com.scalar.db.common.ScannerIterator;
+import com.scalar.db.common.AbstractScanner;
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import java.util.Collections;
-import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
@@ -17,12 +15,11 @@ import software.amazon.awssdk.services.dynamodb.model.GetItemRequest;
 import software.amazon.awssdk.services.dynamodb.model.GetItemResponse;
 
 @NotThreadSafe
-public final class GetItemScanner implements Scanner {
+public final class GetItemScanner extends AbstractScanner {
   private final Map<String, AttributeValue> item;
   private final ResultInterpreter resultInterpreter;
 
   private boolean hasNext;
-  private ScannerIterator scannerIterator;
 
   @SuppressFBWarnings("EI_EXPOSE_REP2")
   public GetItemScanner(
@@ -59,15 +56,6 @@ public final class GetItemScanner implements Scanner {
     Result result = resultInterpreter.interpret(item);
     hasNext = false;
     return Collections.singletonList(result);
-  }
-
-  @Override
-  @Nonnull
-  public Iterator<Result> iterator() {
-    if (scannerIterator == null) {
-      scannerIterator = new ScannerIterator(this);
-    }
-    return scannerIterator;
   }
 
   @Override
